backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

python-dotenv==1.0.1

motor==3.6.0

elasticsearch==8.12.0
aiohttp==3.9.5

openai==1.59.8
qdrant-client==1.13.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

requests==2.32.3
email-validator==2.2.0